Rename the file to match the module.
