+++ lib/dbus.rb
@@ -14,6 +14,7 @@ require_relative "dbus/api_options"
 require_relative "dbus/auth"
 require_relative "dbus/bus"
 require_relative "dbus/bus_name"
+require_relative "dbus/data"
 require_relative "dbus/error"
 require_relative "dbus/introspect"
 require_relative "dbus/logger"
@@ -28,7 +29,6 @@ require_relative "dbus/proxy_object_factory"
 require_relative "dbus/proxy_object_interface"
 require_relative "dbus/raw_message"
 require_relative "dbus/type"
-require_relative "dbus/value"
 require_relative "dbus/xml"
 
 require "socket"
