+++ lib/dbus/data.rb
